move s3 bucket fixture
